cluser options

+++ Cryolo_relion3.0/CryoloExternalJob.py
@@ -96,15 +96,18 @@ def run_job(project_dir, job_dir, args_list):
 
     print(" RELION_IT: Running from model {}".format(model))
 
-    ###CLUSTER
-    os.system(
-        f"{qsub_file} cryolo_predict.py -c config.json -i {os.path.join(project_dir, job_dir, 'cryolo_input')} -o {os.path.join(project_dir, job_dir, 'gen_pick')} -w {model} -g 0 -t {thresh}"
-    )
-    ### WAIT FOR DONEFILE! ###
-    while not os.path.exists(".cry_predict_done"):
-        time.sleep(1)
-    os.remove(".cry_predict_done")
-    ###CLUSTER
+    if relion_it_config.use_cluster:
+        os.system(
+            f"{qsub_file} cryolo_predict.py -c config.json -i {os.path.join(project_dir, job_dir, 'cryolo_input')} -o {os.path.join(project_dir, job_dir, 'gen_pick')} -w {model} -g 0 -t {thresh}"
+        )
+        ### WAIT FOR DONEFILE! ###
+        while not os.path.exists(".cry_predict_done"):
+            time.sleep(1)
+        os.remove(".cry_predict_done")
+    else:
+        os.system(
+            f"cryolo_predict.py -c config.json -i {os.path.join(project_dir, job_dir, 'cryolo_input')} -o {os.path.join(project_dir, job_dir, 'gen_pick')} -w {model} -g 0 -t {thresh}"
+        )
 
     try:
         os.mkdir("picked_stars")
